requirements.txt
numpy~=1.23.4
Pillow~=9.3.0
python-dotenv~=0.21.0
discord.py~=2.3.2